pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "vh-video-container"
version = "1.0.1"
description = "SQLite-based video container format optimized for AI workloads"
readme = "README.md"
license = "MIT"
requires-python = ">=3.8"
authors = [
    {name = "VH System Creative Studio"},
]
keywords = ["video", "container", "sqlite", "ai", "machine-learning", "computer-vision"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Multimedia :: Video",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "numpy",
    "Pillow",
]

[project.optional-dependencies]
viewer = ["Pillow"]
generate = ["torch", "diffusers>=0.25.0", "transformers", "accelerate", "safetensors", "Pillow"]
generate-svd = ["torch", "diffusers>=0.25.0", "transformers", "accelerate", "safetensors", "Pillow"]
generate-kling = ["PyJWT", "requests", "Pillow"]

[project.urls]
Homepage = "https://github.com/titorecbr/vh-container-video-format"
Repository = "https://github.com/titorecbr/vh-container-video-format"

[project.scripts]
vh = "vh_video_container.cli:main"

[tool.setuptools.packages.find]
include = ["vh_video_container*"]
